Move account sensitize-related methods to concern
